feat: wip - add subcommand
